requirements.txt
autopep8==1.3.5
certifi==2018.1.18
chardet==3.0.4
cytoolz==0.9.0
decorator==4.2.1
Django==2.0.2
django-cors-headers==2.2.0
django-extensions==2.0.6
django-mathfilters==0.4.0
djangorestframework==3.7.7
eth-abi==0.5.0
eth-keys==0.1.0b4
eth-tester==0.1.0b11
eth-utils==0.8.1
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
lru-dict==1.1.6
numpy==1.15.0
opencv-python==3.4.2.17
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
py-solc==2.1.0
pycodestyle==2.4.0
Pygments==2.2.0
pysha3==1.0.2
pytz==2018.3
requests==2.18.4
rlp==0.6.0
scipy==1.1.0
semantic-version==2.6.0
simplegeneric==0.8.1
six==1.11.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
web3==3.16.5
